refactor FSDP specific code
